requirements.txt
gTTS==2.5.3
httpx==0.27.2
openai==1.51.0
pymupdf==1.24.11
python-dotenv==1.0.1
python_docx==1.1.2
PyYAML==6.0.2
PyYAML==6.0.2
requests==2.32.3
streamlit==1.38.0
streamlit_authenticator==0.4.1
